starlark: remove custom build context
Now that the `build()` function lives on the concrete type and has
access to the domain specific context, we no longer need a generic
`BuildContext` type to hold PyOxidizer state.

This commit refactors existing users of it to use
`PyOxidizerEnvironmentContext` directly. We delete the wrapper type
from PyOxidizer because it is unused.

There's likely still potential for cleanup here...